Move retry and rate limiting logic for Amazon bedrock out of embeddings. (#6626)
* move rate limiting etc to broader bedrock client

* initla chat

* remove chat

* fix imports

---------
