+++ crates/llms/src/bedrock/embed/mod.rs
@@ -20,7 +20,10 @@ use crate::bedrock::embed::cohere::{
     CohereConfig, CohereEmbedRequest, CohereEmbedResponse, CohereEmbeddingInputType,
     CohereEmbeddingTruncate, CohereEmbeddingType,
 };
-use crate::bedrock::embed::titan::{TitanConfig, TitanEmbedRequest, TitanEmbedResponse};
+use crate::bedrock::embed::titan::{
+    TITAN_TEXT_EMBED_V2, TitanConfig, TitanEmbedRequest, TitanEmbedResponse,
+};
+
 use crate::embeddings::{Embed, Error as EmbedError, Result as EmbedResult};
 use async_openai::error::{ApiError, OpenAIError};
 use async_openai::types::{
@@ -28,37 +31,16 @@ use async_openai::types::{
     EmbeddingVector,
 };
 use async_trait::async_trait;
-use aws_sdk_bedrockruntime::operation::invoke_model::InvokeModelError;
-use aws_sdk_bedrockruntime::{error::SdkError, primitives::Blob};
 use futures::{StreamExt, stream};
-use governor::clock::DefaultClock;
-use governor::state::InMemoryState;
-use governor::{Quota, RateLimiter};
 use serde::Serialize;
 use serde::de::DeserializeOwned;
 use snafu::ResultExt;
 use std::fmt::Debug;
-use std::num::NonZeroU32;
 use std::sync::Arc;
-use tokio::sync::Semaphore;
 use tracing::warn;
-use util::{
-    RetryError,
-    fibonacci_backoff::{FibonacciBackoff, FibonacciBackoffBuilder},
-    retry,
-};
 
 pub mod cohere;
 pub mod titan;
-
-const TITAN_TEXT_EMBED_V2: &str = "amazon.titan-embed-text-v2:0";
-// Maximum number of concurrently running requests.
-// The overall request rate is controlled by the rate_limiter.
-const DEFAULT_MAX_CONCURRENT_INVOCATIONS: usize = 40;
-
-fn default_retry_strategy() -> FibonacciBackoff {
-    FibonacciBackoffBuilder::new().max_retries(Some(10)).build()
-}
 
 #[derive(Debug, Clone)]
 pub struct BedrockEmbed<Rq, Rsp>
@@ -68,13 +50,6 @@ where
 {
     client: BedrockClient,
     config: Arc<dyn BedrockEmbeddingConfig<Rq, Rsp> + 'static>,
-    rate_limiter: Arc<RateLimiter<governor::state::NotKeyed, InMemoryState, DefaultClock>>,
-    // Control the max number of concurrent requests
-    semaphore: Arc<Semaphore>,
-    // Retry strategy for transient or throttling errors
-    retry_strategy: FibonacciBackoff,
-    // Rate limiting configuration for logging and metrics
-    rate_config: BedrockRateLimitConfig,
 }
 
 #[must_use]
@@ -82,10 +57,10 @@ pub fn new_titan_v2(
     client: BedrockClient,
     normalize: bool,
     dimensions: u32,
-    rate_config: BedrockRateLimitConfig,
 ) -> BedrockEmbed<TitanEmbedRequest, TitanEmbedResponse> {
     tracing::debug!(
-        "Initializing Titan v2 embedder: normalize={normalize}, dimensions={dimensions}, rate_limit={rate_config:?}"
+        "Initializing Titan v2 embedder: normalize={normalize}, dimensions={dimensions}, rate_limit={:?}",
+        client.rate_config
     );
 
     let config = Arc::new(TitanConfig {
@@ -94,16 +69,7 @@ pub fn new_titan_v2(
         dimensions,
     }) as Arc<dyn BedrockEmbeddingConfig<TitanEmbedRequest, TitanEmbedResponse>>;
 
-    let rate_limiter = Arc::new(RateLimiter::direct(rate_config.to_quota()));
-
-    BedrockEmbed::<TitanEmbedRequest, TitanEmbedResponse> {
-        client,
-        config,
-        rate_limiter,
-        semaphore: Arc::new(Semaphore::new(rate_config.max_concurrent_invocations)),
-        retry_strategy: default_retry_strategy(),
-        rate_config,
-    }
+    BedrockEmbed::<TitanEmbedRequest, TitanEmbedResponse> { client, config }
 }
 
 #[must_use]
@@ -113,10 +79,10 @@ pub fn new_cohere(
     truncate: CohereEmbeddingTruncate,
     input_type: CohereEmbeddingInputType,
     embedding_type: CohereEmbeddingType,
-    rate_config: BedrockRateLimitConfig,
 ) -> BedrockEmbed<CohereEmbedRequest, CohereEmbedResponse> {
     tracing::debug!(
-        "Initializing Cohere embedder: model_name={model_name}, truncate={truncate:?}, input_type={input_type}, embedding_type={embedding_type}, rate_limit={rate_config:?}"
+        "Initializing Cohere embedder: model_name={model_name}, truncate={truncate:?}, input_type={input_type}, embedding_type={embedding_type}, rate_limit={:?}",
+        client.rate_config
     );
 
     let config = Arc::new(CohereConfig {
@@ -126,16 +92,7 @@ pub fn new_cohere(
         embedding_type,
     }) as Arc<dyn BedrockEmbeddingConfig<CohereEmbedRequest, CohereEmbedResponse>>;
 
-    let rate_limiter = Arc::new(RateLimiter::direct(rate_config.to_quota()));
-
-    BedrockEmbed::<CohereEmbedRequest, CohereEmbedResponse> {
-        client,
-        config,
-        rate_limiter,
-        semaphore: Arc::new(Semaphore::new(rate_config.max_concurrent_invocations)),
-        retry_strategy: default_retry_strategy(),
-        rate_config,
-    }
+    BedrockEmbed::<CohereEmbedRequest, CohereEmbedResponse> { client, config }
 }
 
 impl<Rq, Rsp> BedrockEmbed<Rq, Rsp>
@@ -154,12 +111,11 @@ where
         let mut total_tokens = 0;
 
         // Run embedding requests with up to 5 requests in parallel
-        let mut stream = stream::iter(request_payloads)
-            .map(|req| self.process_single_request(req))
-            .buffered(self.rate_config.max_concurrent_invocations);
+        let mut stream_of_futures =
+            stream::iter(request_payloads).map(|req| self.process_single_request(req));
 
-        while let Some(result) = stream.next().await {
-            let (mut vectors, tokens) = result?;
+        while let Some(result) = stream_of_futures.next().await {
+            let (mut vectors, tokens) = result.await?;
             results.append(&mut vectors);
             total_tokens += tokens;
         }
@@ -177,57 +133,18 @@ where
             })
         })?;
 
-        // Control num concurrent requests
-        let _permit = self.semaphore.acquire().await.map_err(|e| {
-            OpenAIError::ApiError(ApiError {
-                message: format!("Unable to acquire rate limiter permit: {e}"),
-                r#type: None,
-                param: None,
-                code: None,
-            })
-        })?;
-
-        let response = retry(self.retry_strategy.clone(), || async {
-            // Wait for rate limiter
-            self.rate_limiter.until_ready().await;
-
-            match self
-                .client
-                .client
-                .invoke_model()
-                .model_id(self.config.model_id())
-                .body(Blob::new(body.as_bytes()))
-                .content_type("application/json")
-                .send()
-                .await
-            {
-                Ok(response) => Ok(response),
-                Err(e) => Err(match &e {
-                    SdkError::ServiceError(service_error) => match service_error.err() {
-                        InvokeModelError::ThrottlingException(_) => {
-                            tracing::debug!(
-                                "Bedrock embedding model throttled, backing off and retrying..."
-                            );
-                            RetryError::transient(e)
-                        }
-                        _ => RetryError::permanent(e),
-                    },
-                    _ => RetryError::permanent(e),
-                }),
-            }
-        })
-        .await
-        .map_err(|e| {
-            OpenAIError::ApiError(ApiError {
-                message: match e.into_source() {
-                    Ok(s_err) => s_err.to_string(),
-                    Err(e) => e.to_string(),
-                },
-                r#type: None,
-                param: None,
-                code: None,
-            })
-        })?;
+        let response = self
+            .client
+            .do_the_thing(self.config.model_id().clone(), body)
+            .await
+            .map_err(|e| {
+                OpenAIError::ApiError(ApiError {
+                    message: e.to_string(),
+                    r#type: None,
+                    param: None,
+                    code: None,
+                })
+            })?;
 
         let response_body = response.body().as_ref();
         let response_obj = serde_json::from_slice(response_body).boxed().map_err(|e| {
@@ -333,13 +250,6 @@ where
         let texts = Self::convert_input_to_texts(&input);
 
         let num_items = texts.len();
-        tracing::trace!(
-            "Embedding {} records using model {} (max_concurrent_invocations: {}, requests_per_minute_limit: {})",
-            num_items,
-            self.config.model_id(),
-            self.rate_config.max_concurrent_invocations,
-            self.rate_config.requests_per_minute_limit
-        );
 
         let start = std::time::Instant::now();
 
@@ -364,68 +274,5 @@ where
 
     fn size(&self) -> i32 {
         self.config.dimensions()
-    }
-}
-
-#[derive(Debug)]
-pub struct BedrockRateLimitConfigBuilder {
-    requests_per_minute_limit: Option<u32>,
-    max_concurrent_invocations: Option<usize>,
-}
-
-impl Default for BedrockRateLimitConfigBuilder {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl BedrockRateLimitConfigBuilder {
-    #[must_use]
-    pub fn new() -> Self {
-        Self {
-            requests_per_minute_limit: None,
-            max_concurrent_invocations: None,
-        }
-    }
-
-    #[must_use]
-    pub fn requests_per_minute(&mut self, limit: u32) -> &Self {
-        self.requests_per_minute_limit = Some(limit);
-        self
-    }
-
-    #[must_use]
-    pub fn max_concurrent_invocations(&mut self, limit: usize) -> &Self {
-        self.max_concurrent_invocations = Some(limit);
-        self
-    }
-
-    #[must_use]
-    pub fn build(self) -> BedrockRateLimitConfig {
-        BedrockRateLimitConfig {
-            requests_per_minute_limit: self.requests_per_minute_limit.unwrap_or(1_500),
-            max_concurrent_invocations: self
-                .max_concurrent_invocations
-                .unwrap_or(DEFAULT_MAX_CONCURRENT_INVOCATIONS),
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct BedrockRateLimitConfig {
-    pub requests_per_minute_limit: u32,
-    pub max_concurrent_invocations: usize,
-}
-
-impl BedrockRateLimitConfig {
-    #[must_use]
-    pub fn to_quota(&self) -> Quota {
-        Quota::per_minute(
-            NonZeroU32::new(self.requests_per_minute_limit).unwrap_or_else(|| {
-                unreachable!(
-                    "requests_per_minute_limit is u32 and should always successfully convert to NonZeroU32"
-                )
-            }),
-        )
     }
 }
